requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openpyxl==3.1.5
Pillow==10.4.0
python-multipart==0.0.9
httpx==0.27.0
python-jose[cryptography]==3.3.0
supabase==2.7.4

